Old parser removed and renamed gherkin_parser.py to parser.py
