+++ src/pytest_bdd/parser.py
@@ -1,22 +1,23 @@
-from __future__ import annotations
-
-import os.path
+import linecache
 import re
 import textwrap
-from collections import OrderedDict
-from dataclasses import dataclass, field
-from typing import Any, Dict, Iterable, List, Mapping, Optional, Sequence
+from pathlib import Path
+from typing import Any, Dict, Iterable, List, Mapping, Optional, Tuple, Union
 
-from .gherkin_parser import Background as GherkinBackground
-from .gherkin_parser import Feature as GherkinFeature
-from .gherkin_parser import GherkinDocument, GherkinParser
-from .gherkin_parser import Scenario as GherkinScenario
-from .gherkin_parser import Step as GherkinStep
-from .gherkin_parser import Tag as GherkinTag
-from .types import GIVEN, THEN, WHEN
+from gherkin.errors import CompositeParserException
+from gherkin.parser import Parser
+from gherkin.token_scanner import TokenScanner
+from pydantic import BaseModel, field_validator, model_validator
+
+from . import exceptions
+from .types import STEP_TYPES
 
 STEP_PARAM_RE = re.compile(r"<(.+?)>")
 COMMENT_RE = re.compile(r"(^|(?<=\s))#")
+
+
+def check_instance_by_name(obj: Any, class_name: str) -> bool:
+    return obj.__class__.__name__ == class_name
 
 
 def strip_comments(line: str) -> str:
@@ -33,237 +34,133 @@ def strip_comments(line: str) -> str:
     return line.strip()
 
 
-@dataclass(eq=False)
-class Feature:
-    """Represents a feature parsed from a feature file.
-
-    Attributes:
-        scenarios (OrderedDict[str, ScenarioTemplate]): A dictionary of scenarios in the feature.
-        filename (str): The absolute path of the feature file.
-        rel_filename (str): The relative path of the feature file.
-        name (Optional[str]): The name of the feature.
-        tags (set[str]): A set of tags associated with the feature.
-        background (Optional[Background]): The background steps for the feature, if any.
-        line_number (int): The line number where the feature starts in the file.
-        description (str): The description of the feature.
-    """
-
-    scenarios: OrderedDict[str, ScenarioTemplate]
-    filename: str
-    rel_filename: str
-    name: str | None
-    tags: set[str]
-    background: Background | None
-    line_number: int
-    description: str
+class Location(BaseModel):
+    column: int
+    line: int
 
 
-@dataclass(eq=False)
-class Examples:
-    """Represents examples used in scenarios for parameterization.
+class Comment(BaseModel):
+    location: Location
+    text: str
 
-    Attributes:
-        line_number (Optional[int]): The line number where the examples start.
-        name (Optional[str]): The name of the examples.
-        example_params (List[str]): The names of the parameters for the examples.
-        examples (List[Sequence[str]]): The list of example rows.
-    """
 
-    line_number: int | None = None
-    name: str | None = None
-    example_params: list[str] = field(default_factory=list)
-    examples: list[Sequence[str]] = field(default_factory=list)
+class Cell(BaseModel):
+    location: Location
+    value: str
 
-    def set_param_names(self, keys: Iterable[str]) -> None:
-        """Set the parameter names for the examples.
 
-        Args:
-            keys (Iterable[str]): The parameter names to set.
+class Row(BaseModel):
+    id: str
+    location: Location
+    cells: List[Cell]
+
+
+class DataTable(BaseModel):
+    name: Optional[str] = None
+    location: Location
+    tableHeader: Optional[Row] = None
+    tableBody: Optional[List[Row]] = None
+
+    def as_contexts(self) -> Iterable[Dict[str, Any]]:
         """
-        self.example_params = [str(key) for key in keys]
-
-    def add_example(self, values: Sequence[str]) -> None:
-        """Add a new example row.
-
-        Args:
-            values (Sequence[str]): The values for the example row.
-        """
-        self.examples.append([str(value) if value is not None else "" for value in values])
-
-    def as_contexts(self) -> Iterable[dict[str, Any]]:
-        """Generate contexts for the examples.
+        Generate contexts for the examples.
 
         Yields:
             Dict[str, Any]: A dictionary mapping parameter names to their values for each example row.
         """
-        for row in self.examples:
-            assert len(self.example_params) == len(row)
-            yield dict(zip(self.example_params, row))
+        if not self.tableHeader or not self.tableBody:
+            return  # If header or body is missing, there's nothing to yield
 
-    def __bool__(self) -> bool:
-        """Check if there are any examples.
+        # Extract parameter names from the tableHeader (row with headers)
+        example_params = [cell.value for cell in self.tableHeader.cells]
 
-        Returns:
-            bool: True if there are examples, False otherwise.
-        """
-        return bool(self.examples)
+        for row in self.tableBody:
+            assert len(example_params) == len(row.cells), "Row length does not match header length"
+            # Map parameter names (from header) to values (from the row)
+            yield dict(zip(example_params, [cell.value for cell in row.cells]))
 
 
-@dataclass(eq=False)
-class ScenarioTemplate:
-    """Represents a scenario template within a feature.
+class DocString(BaseModel):
+    content: str
+    delimiter: str
+    location: Location
 
-    Attributes:
-        feature (Feature): The feature to which this scenario belongs.
-        name (str): The name of the scenario.
-        line_number (int): The line number where the scenario starts in the file.
-        templated (bool): Whether the scenario is templated.
-        description (Optional[str]): The description of the scenario.
-        tags (set[str]): A set of tags associated with the scenario.
-        _steps (List[Step]): The list of steps in the scenario (internal use only).
-        examples (Optional[Examples]): The examples used for parameterization in the scenario.
-    """
+    @field_validator("content", mode="before")
+    def dedent_content(cls, value: str) -> str:
+        return textwrap.dedent(value)
 
-    feature: Feature
-    name: str
-    line_number: int
-    templated: bool
-    description: str | None = None
-    tags: set[str] = field(default_factory=set)
-    _steps: list[Step] = field(init=False, default_factory=list)
-    examples: Examples | None = field(default_factory=Examples)
 
-    def add_step(self, step: Step) -> None:
-        """Add a step to the scenario.
-
-        Args:
-            step (Step): The step to add.
-        """
-        step.scenario = self
-        self._steps.append(step)
+class Step(BaseModel):
+    id: str
+    keyword: str
+    keywordType: str
+    location: Location
+    text: str
+    dataTable: Optional[DataTable] = None
+    docString: Optional[DocString] = None
+    raw_name: Optional[str] = None
+    name: Optional[str] = None
+    parent: Optional[Union["Background", "Scenario"]] = None
+    failed: bool = False
+    duration: Optional[float] = None
 
     @property
-    def steps(self) -> list[Step]:
-        """Get all steps for the scenario, including background steps.
+    def scenario(self) -> Optional["Scenario"]:
+        """Returns the scenario if the step's parent is a Scenario."""
+        if isinstance(self.parent, Scenario):
+            return self.parent
+        return None
 
-        Returns:
-            List[Step]: A list of steps, including any background steps from the feature.
-        """
-        return (self.feature.background.steps if self.feature.background else []) + self._steps
+    @property
+    def background(self) -> Optional["Background"]:
+        """Returns the background if the step's parent is a Background."""
+        if isinstance(self.parent, Background):
+            return self.parent
+        return None
 
-    def render(self, context: Mapping[str, Any]) -> Scenario:
-        """Render the scenario with the given context.
+    def generate_initial_name(self) -> None:
+        """Generate an initial name based on the step's text and optional docString."""
+        self.name = strip_comments(self.text)
+        if self.docString:
+            self.name = f"{self.name}\n{self.docString.content}"
+        # Populate a frozen copy of the name untouched by params later
+        self.raw_name = self.name
 
-        Args:
-            context (Mapping[str, Any]): The context for rendering steps.
+    @model_validator(mode="after")
+    def set_name(cls, instance):
+        """Set the 'name' attribute after model validation if it is not already provided."""
+        instance.generate_initial_name()
+        return instance
 
-        Returns:
-            Scenario: A Scenario object with steps rendered based on the context.
-        """
-        background_steps = self.feature.background.steps if self.feature.background else []
-        scenario_steps = [
-            Step(
-                name=step.render(context),
-                type=step.type,
-                indent=step.indent,
-                line_number=step.line_number,
-                keyword=step.keyword,
-            )
-            for step in self._steps
-        ]
-        steps = background_steps + scenario_steps
-        return Scenario(
-            feature=self.feature,
-            name=self.name,
-            line_number=self.line_number,
-            steps=steps,
-            tags=self.tags,
-            description=self.description,
-        )
+    @field_validator("keyword", mode="before")
+    def normalize_keyword(cls, value: str) -> str:
+        """Normalize the keyword (e.g., Given, When, Then)."""
+        return value.title().strip()
 
+    @property
+    def given_when_then(self) -> str:
+        """Get the Given/When/Then form of the step."""
+        return self._gwt
 
-@dataclass(eq=False)
-class Scenario:
-    """Represents a scenario with steps.
-
-    Attributes:
-        feature (Feature): The feature to which this scenario belongs.
-        name (str): The name of the scenario.
-        line_number (int): The line number where the scenario starts in the file.
-        steps (List[Step]): The list of steps in the scenario.
-        description (Optional[str]): The description of the scenario.
-        tags (set[str]): A set of tags associated with the scenario.
-    """
-
-    feature: Feature
-    name: str
-    line_number: int
-    steps: list[Step]
-    description: str | None = None
-    tags: set[str] = field(default_factory=set)
-
-
-@dataclass(eq=False)
-class Step:
-    """Represents a step within a scenario or background.
-
-    Attributes:
-        type (str): The type of step (e.g., 'given', 'when', 'then').
-        name (str): The name of the step.
-        line_number (int): The line number where the step starts in the file.
-        indent (int): The indentation level of the step.
-        keyword (str): The keyword used for the step (e.g., 'Given', 'When', 'Then').
-        failed (bool): Whether the step has failed (internal use only).
-    """
-
-    type: str
-    name: str
-    line_number: int
-    indent: int
-    keyword: str
-    failed: bool = field(init=False, default=False)
-
-    def __init__(self, name: str, type: str, indent: int, line_number: int, keyword: str) -> None:
-        """Initialize a step.
-
-        Args:
-            name (str): The name of the step.
-            type (str): The type of the step (e.g., 'given', 'when', 'then').
-            indent (int): The indentation level of the step.
-            line_number (int): The line number where the step starts in the file.
-            keyword (str): The keyword used for the step (e.g., 'Given', 'When', 'Then').
-        """
-        self.name = name
-        self.type = type
-        self.indent = indent
-        self.line_number = line_number
-        self.keyword = keyword
+    @given_when_then.setter
+    def given_when_then(self, gwt: str) -> None:
+        """Set the Given/When/Then form of the step."""
+        self._gwt = gwt
 
     def __str__(self) -> str:
-        """Return a string representation of the step.
-
-        Returns:
-            str: A string representation of the step.
-        """
-        return f'{self.type.capitalize()} "{self.name}"'
+        """Return a string representation of the step."""
+        return f'{self.given_when_then.capitalize()} "{self.name}"'
 
     @property
-    def params(self) -> tuple[str, ...]:
-        """Get the parameters in the step name.
+    def params(self) -> Tuple[str, ...]:
+        """Get the parameters in the step name."""
+        return tuple(frozenset(STEP_PARAM_RE.findall(self.raw_name)))
 
-        Returns:
-            Tuple[str, ...]: A tuple of parameter names found in the step name.
-        """
-        return tuple(frozenset(STEP_PARAM_RE.findall(self.name)))
-
-    def render(self, context: Mapping[str, Any]) -> str:
-        """Render the step name with the given context, but avoid replacing text inside angle brackets if context is missing.
+    def render(self, context: Mapping[str, Any]) -> None:
+        """Render the step name with the given context and update the instance.
 
         Args:
             context (Mapping[str, Any]): The context for rendering the step name.
-
-        Returns:
-            str: The rendered step name with parameters replaced only if they exist in the context.
         """
 
         def replacer(m: re.Match) -> str:
@@ -271,173 +168,261 @@ class Step:
             # If the context contains the variable, replace it. Otherwise, leave it unchanged.
             return str(context.get(varname, f"<{varname}>"))
 
-        return STEP_PARAM_RE.sub(replacer, self.name)
+        # Render the name and update the instance's text attribute
+        rendered_name = STEP_PARAM_RE.sub(replacer, self.raw_name)
+        self.name = rendered_name
 
 
-@dataclass(eq=False)
-class Background:
-    """Represents the background steps for a feature.
+class Tag(BaseModel):
+    id: str
+    location: Location
+    name: str
 
-    Attributes:
-        feature (Feature): The feature to which this background belongs.
-        line_number (int): The line number where the background starts in the file.
-        steps (List[Step]): The list of steps in the background.
-    """
 
-    feature: Feature
-    line_number: int
-    steps: list[Step] = field(init=False, default_factory=list)
+class Scenario(BaseModel):
+    id: str
+    keyword: str
+    location: Location
+    name: str
+    description: str
+    steps: List[Step]
+    tags: List[Tag]
+    examples: Optional[List[DataTable]] = None
+    parent: Optional[Union["Feature", "Rule"]] = None
 
-    def add_step(self, step: Step) -> None:
-        """Add a step to the background.
+    @field_validator("description", mode="before")
+    def dedent_description(cls, value: str) -> str:
+        return textwrap.dedent(value)
+
+    @model_validator(mode="after")
+    def process_steps(cls, instance):
+        steps = instance.steps
+        instance.steps = _compute_given_when_then(steps)
+        return instance
+
+    @model_validator(mode="after")
+    def process_scenario_for_steps(cls, instance):
+        for step in instance.steps:
+            step.parent = instance
+        return instance
+
+    @property
+    def tag_names(self) -> List[str]:
+        return get_tag_names(self.tags)
+
+    def render(self, context: Mapping[str, Any]) -> None:
+        """Render the scenario's steps with the given context.
 
         Args:
-            step (Step): The step to add.
+            context (Mapping[str, Any]): The context for rendering steps.
         """
-        step.background = self
-        self.steps.append(step)
+        for step in self.steps:
+            step.render(context)
+
+    @property
+    def feature(self):
+        if check_instance_by_name(self.parent, "Feature"):
+            return self.parent
+        return None
+
+    @property
+    def rule(self):
+        if check_instance_by_name(self.parent, "Rule"):
+            return self.parent
+        return None
+
+    @property
+    def all_steps(self) -> List[Step]:
+        """Get all steps including background steps if present."""
+        # Check if the scenario belongs to a feature and if the feature has background steps
+        background_steps = self.feature.background_steps if self.feature else []
+        # Return the combined list of background steps and scenario steps
+        return background_steps + self.steps
 
 
-class FeatureParser:
-    """Converts a feature file into a Feature object.
+class Rule(BaseModel):
+    id: str
+    keyword: str
+    location: Location
+    name: str
+    description: str
+    tags: List[Tag]
+    children: List[Scenario]
+    parent: Optional["Feature"] = None
 
-    Args:
-        basedir (str): The basedir for locating feature files.
-        filename (str): The filename of the feature file.
-        encoding (str): File encoding of the feature file to parse.
-    """
+    @field_validator("description", mode="before")
+    def dedent_description(cls, value: str) -> str:
+        return textwrap.dedent(value)
 
-    def __init__(self, basedir: str, filename: str, encoding: str = "utf-8"):
-        self.abs_filename = os.path.abspath(os.path.join(basedir, filename))
-        self.rel_filename = os.path.join(os.path.basename(basedir), filename)
+    @model_validator(mode="after")
+    def process_scenarios(cls, instance):
+        for scenario in instance.children:
+            scenario.parent = instance
+        return instance
+
+    @property
+    def tag_names(self) -> List[str]:
+        return get_tag_names(self.tags)
+
+
+class Background(BaseModel):
+    id: str
+    keyword: str
+    location: Location
+    name: str
+    description: str
+    steps: List[Step]
+    parent: Optional["Feature"] = None
+
+    @field_validator("description", mode="before")
+    def dedent_description(cls, value: str) -> str:
+        return textwrap.dedent(value)
+
+    @model_validator(mode="after")
+    def process_given_when_then(cls, instance):
+        steps = instance.steps
+        instance.steps = _compute_given_when_then(steps)
+        return instance
+
+    @model_validator(mode="after")
+    def process_background_for_steps(cls, instance):
+        for step in instance.steps:
+            step.parent = instance
+        return instance
+
+    def render(self, context: Mapping[str, Any]) -> None:
+        """Render the scenario's steps with the given context.
+
+        Args:
+            context (Mapping[str, Any]): The context for rendering steps.
+        """
+        for step in self.steps:
+            step.render(context)
+
+
+class Child(BaseModel):
+    background: Optional[Background] = None
+    rule: Optional[Rule] = None
+    scenario: Optional[Scenario] = None
+    parent: Optional[Union["Feature", "Rule"]] = None
+
+    @model_validator(mode="after")
+    def assign_parents(cls, instance):
+        if instance.scenario:
+            instance.scenario.parent = instance.parent
+        if instance.background:
+            instance.background.parent = instance.parent
+        return instance
+
+
+class Feature(BaseModel):
+    keyword: str
+    location: Location
+    tags: List[Tag]
+    name: str
+    description: str
+    children: List[Child]
+    abs_filename: Optional[str] = None
+    rel_filename: Optional[str] = None
+
+    @field_validator("description", mode="before")
+    def dedent_description(cls, value: str) -> str:
+        return textwrap.dedent(value)
+
+    @model_validator(mode="after")
+    def assign_child_parents(cls, instance):
+        for child in instance.children:
+            child.parent = instance
+            if child.scenario:
+                child.scenario.parent = instance
+            if child.background:
+                child.background.parent = instance
+        return instance
+
+    @property
+    def filename(self) -> Optional[str]:
+        """
+        Returns the file name from abs_filename, if available.
+        """
+        if self.abs_filename:
+            return str(Path(self.abs_filename).resolve())
+        return None
+
+    @property
+    def scenarios(self) -> List[Scenario]:
+        return [child.scenario for child in self.children if child.scenario]
+
+    @property
+    def backgrounds(self) -> List[Background]:
+        return [child.background for child in self.children if child.background]
+
+    @property
+    def background_steps(self) -> List[Step]:
+        _steps = []
+        for background in self.backgrounds:
+            _steps.extend(background.steps)
+        return _steps
+
+    @property
+    def rules(self) -> List[Rule]:
+        return [child.rule for child in self.children if child.rule]
+
+    def get_child_by_name(self, name: str) -> Optional[Union[Scenario, Background]]:
+        """
+        Returns the child (Scenario or Background) that has the given name.
+        """
+        for scenario in self.scenarios:
+            if scenario.name == name:
+                return scenario
+        for background in self.backgrounds:
+            if background.name == name:
+                return background
+        return None
+
+    @property
+    def tag_names(self) -> List[str]:
+        return get_tag_names(self.tags)
+
+
+class GherkinDocument(BaseModel):
+    feature: Feature
+    comments: List[Comment]
+
+
+def _compute_given_when_then(steps: List[Step]) -> List[Step]:
+    last_gwt = None
+    for step in steps:
+        if step.keyword.lower() in STEP_TYPES:
+            last_gwt = step.keyword.lower()
+        step.given_when_then = last_gwt
+    return steps
+
+
+def get_tag_names(tags: List[Tag]):
+    return [tag.name.lstrip("@") for tag in tags]
+
+
+class GherkinParser:
+    def __init__(self, abs_filename: str, rel_filename: str, encoding: str = "utf-8"):
+        self.abs_filename = abs_filename
+        self.rel_filename = rel_filename
         self.encoding = encoding
 
-    @staticmethod
-    def get_tag_names(tag_data: list[GherkinTag]) -> set[str]:
-        """Extract tag names from tag data.
+        with open(self.abs_filename, encoding=self.encoding) as f:
+            self.feature_file_text = f.read()
+        try:
+            self.gherkin_data = Parser().parse(TokenScanner(self.feature_file_text))
+        except CompositeParserException as e:
+            raise exceptions.FeatureError(
+                e.args[0],
+                e.errors[0].location["line"],
+                linecache.getline(self.abs_filename, e.errors[0].location["line"]).rstrip("\n"),
+                self.abs_filename,
+            ) from e
 
-        Args:
-            tag_data (List[dict]): The tag data to extract names from.
-
-        Returns:
-            set[str]: A set of tag names.
-        """
-        return {tag.name.lstrip("@") for tag in tag_data}
-
-    @staticmethod
-    def get_step_type(keyword: str) -> str | None:
-        """Map a step keyword to its corresponding type.
-
-        Args:
-            keyword (str): The keyword for the step (e.g., 'given', 'when', 'then').
-
-        Returns:
-            Optional[str]: The type of the step, or None if the keyword is unknown.
-        """
-        return {
-            "given": GIVEN,
-            "when": WHEN,
-            "then": THEN,
-        }.get(keyword)
-
-    @staticmethod
-    def parse_steps(steps_data: list[GherkinStep]) -> list[Step]:
-        """Parse a list of step data into Step objects.
-
-        Args:
-            steps_data (List[dict]): The list of step data.
-
-        Returns:
-            List[Step]: A list of Step objects.
-        """
-        steps = []
-        for step_data in steps_data:
-            name = strip_comments(step_data.text)
-            if step_data.docString:
-                name = f"{name}\n{step_data.docString.content}"
-            steps.append(
-                Step(
-                    name=name,
-                    type=step_data.given_when_then,
-                    indent=step_data.location.column - 1,
-                    line_number=step_data.location.line,
-                    keyword=step_data.keyword.title(),
-                )
-            )
-        return steps
-
-    def parse_scenario(self, scenario_data: GherkinScenario, feature: Feature) -> ScenarioTemplate:
-        """Parse a scenario data dictionary into a ScenarioTemplate object.
-
-        Args:
-            scenario_data (dict): The dictionary containing scenario data.
-            feature (Feature): The feature to which this scenario belongs.
-
-        Returns:
-            ScenarioTemplate: A ScenarioTemplate object representing the parsed scenario.
-        """
-        templated = bool(scenario_data.examples)
-        scenario = ScenarioTemplate(
-            feature=feature,
-            name=strip_comments(scenario_data.name),
-            line_number=scenario_data.location.line,
-            templated=templated,
-            tags=self.get_tag_names(scenario_data.tags),
-            description=textwrap.dedent(scenario_data.description),
-        )
-        for step in self.parse_steps(scenario_data.steps):
-            scenario.add_step(step)
-
-        for example_data in scenario_data.examples:
-            examples = Examples(
-                line_number=example_data.location.line,
-                name=example_data.name,
-            )
-            param_names = [cell.value for cell in example_data.tableHeader.cells]
-            examples.set_param_names(param_names)
-            for row in example_data.tableBody:
-                values = [cell.value or "" for cell in row.cells]
-                examples.add_example(values)
-            scenario.examples = examples
-
-        return scenario
-
-    def parse_background(self, background_data: GherkinBackground, feature: Feature) -> Background:
-        background = Background(
-            feature=feature,
-            line_number=background_data.location.line,
-        )
-        background.steps = self.parse_steps(background_data.steps)
-        return background
-
-    def _parse_feature_file(self) -> GherkinDocument:
-        """Parse a feature file into a Feature object.
-
-        Returns:
-            Dict: A Gherkin document representation of the feature file.
-        """
-        return GherkinParser(self.abs_filename, self.rel_filename, self.encoding).to_gherkin_document()
-
-    def parse(self):
-        gherkin_doc: GherkinDocument = self._parse_feature_file()
-        feature_data: GherkinFeature = gherkin_doc.feature
-        return feature_data
-        # feature = Feature(
-        #     scenarios=OrderedDict(),
-        #     filename=self.abs_filename,
-        #     rel_filename=self.rel_filename,
-        #     name=strip_comments(feature_data.name),
-        #     tags=self.get_tag_names(feature_data.tags),
-        #     background=None,
-        #     line_number=feature_data.location.line,
-        #     description=textwrap.dedent(feature_data.description),
-        # )
-        #
-        # for child in feature_data.children:
-        #     if child.background:
-        #         feature.background = self.parse_background(child.background, feature)
-        #     elif child.scenario:
-        #         scenario = self.parse_scenario(child.scenario, feature)
-        #         feature.scenarios[scenario.name] = scenario
-        #
-        # return feature
+    def to_gherkin_document(self) -> GherkinDocument:
+        gherkin_document = GherkinDocument(**self.gherkin_data)
+        # Pass abs_filename to the feature
+        gherkin_document.feature.abs_filename = self.abs_filename
+        gherkin_document.feature.rel_filename = self.rel_filename
+        return gherkin_document
