+++ src/pytest_bdd/reporting.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from _pytest.reports import TestReport
     from _pytest.runner import CallInfo
 
-    from .gherkin_parser import Feature, Scenario, Step
+    from .parser import Feature, Scenario, Step
 
 
 class StepReport:
@@ -125,7 +125,8 @@ class ScenarioReport:
     def fail(self) -> None:
         """Stop collecting information and finalize the report as failed."""
         self.current_step_report.finalize(failed=True)
-        remaining_steps = self.scenario.steps[len(self.step_reports) :]
+        steps = self.scenario.all_steps
+        remaining_steps = steps[len(self.step_reports) :]
 
         # Fail the rest of the steps and make reports.
         for step in remaining_steps:
